More codepath simplification, convert lru to track u32

+++ code/font/vefontcache/vefontcache.odin
@@ -40,7 +40,7 @@ Context :: struct {
 	entries : [dynamic]Entry,
 
 	temp_path               : [dynamic]Vertex,
-	temp_codepoint_seen     : map[u64]b8,
+	temp_codepoint_seen     : map[u32]b8,
 	temp_codepoint_seen_num : i32,
 
 	snap_width  : f32,
@@ -183,7 +183,7 @@ startup :: proc( ctx : ^Context, parser_kind : Parser_Kind = .STB_TrueType,
 	temp_path, error = make( [dynamic]Vertex, len = 0, cap = temp_path_reserve )
 	assert(error == .None, "VEFontCache.init : Failed to allocate temp_path")
 
-	temp_codepoint_seen, error = make( map[u64]b8, uint(temp_codepoint_seen_reserve) )
+	temp_codepoint_seen, error = make( map[u32]b8, uint(temp_codepoint_seen_reserve) )
 	assert(error == .None, "VEFontCache.init : Failed to allocate temp_path")
 
 	draw_list.vertices, error = make( [dynamic]Vertex, len = 0, cap = 8 * Kilobyte )
